+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Changed `res` argument type in `ApiSectionUpload.auto` method
   * Changed `file` argument type in `ApiSectionUpload.base` & `ApiSectionUpload.multipart` methods
 * Added ability to upload files in example project in `flutter_web` environment
+* Removed deprecated `detectFaces` method from `ApiFiles`;
 
 ## [1.4.1] - Mon Dec 9 2019
 * Updated dependencies to the latest version